+++ src/interfaces/applications/items/IHaveItem.php
@@ -1,8 +1,0 @@
-<?php
-namespace deflou\interfaces\applications\items;
-
-interface IHaveItem
-{
-    public function getItem(string $name): array;
-    public function getItems(): array;
-}